12. Numbers have life; they're not just symbols on paper

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,7 +9,7 @@
 
 void print_number(int n)
 {
-	unsigned int k;
+	unsigned long int k;
 
 	if (n < 0)
 	{
